sección finalizada

+++ src/todos/app.js
@@ -1,11 +1,14 @@
-import todoStore from "../store/todo.store";
+import todoStore, { Filters } from "../store/todo.store";
 import html  from "./app.html?raw";
-import { renderTodos } from "./use-cases";
+import { renderTodos, renderPending } from "./use-cases";
 
 
 const ElementsIDs = { //esto se hace para evitar poner strings como identificadores
+    ClearCompletedButton : '.clear-completed',
     TodoList: '.todo-list', //ul element
-    NewTodoInput: '#new-todo-input'
+    NewTodoInput: '#new-todo-input',
+    TodoFilters : '.filtro', 
+    PendingCountLabel : '#pending-count',
 
 }
 /**
@@ -19,14 +22,21 @@ export const App = ( elementId ) => {
         const todos = todoStore.getTodos(  todoStore.getCurrentFilter() );
         console.log( todos)
         renderTodos( ElementsIDs.TodoList , todos );
+        updatePendingCount();
     }   
+
+    const updatePendingCount = () => {
+        renderPending( ElementsIDs.PendingCountLabel);
+    }
+
+
     //cuando la funcion App se llama
     (() => {
         const app = document.createElement('div');
         app.innerHTML = html;
         document.querySelector(elementId).append( app );
     //elementId va a decir donde quiero que se renderice eso, el app
-        displayTodos()
+        displayTodos();
     })(); 
     //la funcion autoinbocada se usa aquí porque se necesita
     // la función esté creada y se necesita hacer referencia
@@ -35,9 +45,14 @@ export const App = ( elementId ) => {
 
     //lo siguiente se pone debajo de la funcion autoinbocada
     //porque antes no existe
+
+
     //Referencias HTML
     const newDescriptionInput = document.querySelector( ElementsIDs.NewTodoInput );
     const todoListUL = document.querySelector( ElementsIDs.TodoList);
+    const clearCompletedButton = document.querySelector ( ElementsIDs.ClearCompletedButton );
+    const filtersLIs = document.querySelectorAll(ElementsIDs.TodoFilters ); 
+    
     //Listeners
     newDescriptionInput.addEventListener('keyup', (event) => {
         
@@ -59,10 +74,35 @@ export const App = ( elementId ) => {
     todoListUL.addEventListener('click', (event) => {
         const isDestroyedElement = event.target.className === 'destroy';
         const element = event.target.closest('[data-id]');//el padre mas cercano, el elemento html que tenga ese atribute, no hacia abajo sino hacia el padre
-        if( !isDestroyedElement ) return;
+        if( !element || !isDestroyedElement ) return;
 
         todoStore.deleteTodo(element.getAttribute('data-id'))
         displayTodos();
     });
 
+    clearCompletedButton.addEventListener('click', () => {
+        todoStore.deleteCompleted();
+        displayTodos();
+
+    });
+
+    filtersLIs.forEach( element => {
+        element.addEventListener('click', (element) => { 
+            filtersLIs.forEach( e => e.classList.remove('selected'));   
+            element.target.classList.add('selected');
+
+            switch( element.target.text ){
+                case 'Todos':
+                    todoStore.setFilter( Filters.All)
+                break;
+                case 'Pendientes':
+                    todoStore.setFilter( Filters.Pending)
+                break;
+                case 'Completados':
+                    todoStore.setFilter( Filters.Completed)
+                break;
+            }
+                displayTodos();
+        });
+    });
 }
